修复 compiler-mini 编译器在严格编译选项下的警告：
1. 符号比较警告 (codegen.c:2791, 2822)
   - 原因：ASTNodeType 是无符号枚举，与 -1 的有符号 int 不匹配
   - 解决：添加 (int) 类型转换

2. 未使用变量警告 (parser.c:92-94)
   - 原因：三个变量声明后未使用
   - 解决：移除未使用的变量声明

+++ compiler-mini/src/codegen.c
@@ -2788,7 +2788,7 @@ LLVMValueRef codegen_gen_expr(CodeGenerator *codegen, ASTNode *expr) {
             // 从第一个元素生成值以推断元素类型
             LLVMValueRef first_element_val = codegen_gen_expr(codegen, elements[0]);
             if (!first_element_val) {
-                fprintf(stderr, "错误: 数组字面量第一个元素生成失败 (元素类型: %d)\n", elements[0] ? elements[0]->type : -1);
+                fprintf(stderr, "错误: 数组字面量第一个元素生成失败 (元素类型: %d)\n", elements[0] ? (int)elements[0]->type : -1);
                 return NULL;
             }
             
@@ -2819,7 +2819,7 @@ LLVMValueRef codegen_gen_expr(CodeGenerator *codegen, ASTNode *expr) {
                 // 生成元素值
                 LLVMValueRef element_val = codegen_gen_expr(codegen, element);
                 if (!element_val) {
-                    fprintf(stderr, "错误: 数组字面量第 %d 个元素生成失败 (元素类型: %d)\n", i, element ? element->type : -1);
+                    fprintf(stderr, "错误: 数组字面量第 %d 个元素生成失败 (元素类型: %d)\n", i, element ? (int)element->type : -1);
                     return NULL;
                 }
